feat: refine UI based on authorization and link image to website

+++ src/lib/components/PlacesList.svelte
@@ -4,10 +4,12 @@
   import { db } from "../firebase";
   import Card from "./Card.svelte";
   import Button from "./Button.svelte";
-  import EditPlaceForm from "./EditPlaceForm.svelte"; // Import the EditPlaceForm component
+  import EditPlaceForm from "./EditPlaceForm.svelte";
+
+  export let isAuthorizedToWrite; // Prop to receive authorization status
 
   let places = [];
-  let editingPlace = null; // State to hold the place being edited
+  let editingPlace = null;
 
   onMount(() => {
     const unsubscribe = onSnapshot(collection(db, "places"), (snapshot) => {
@@ -27,15 +29,15 @@
   }
 
   function editPlace(place) {
-    editingPlace = place; // Set the place to be edited
+    editingPlace = place;
   }
 
   function handleEditSave() {
-    editingPlace = null; // Clear editing state after save
+    editingPlace = null;
   }
 
   function handleEditCancel() {
-    editingPlace = null; // Clear editing state if canceled
+    editingPlace = null;
   }
 </script>
 
@@ -55,12 +57,16 @@
           <a href={place.website} target="_blank" rel="noreferrer">Website</a>
         {/if}
         {#if place.image}
-          <img src={place.image} alt={place.name} />
+          <a href={place.website} target="_blank" rel="noreferrer">
+            <img src={place.image} alt={place.name} />
+          </a>
         {/if}
-        <div class="actions">
-          <Button label="Edit" onClick={() => editPlace(place)} />
-          <Button label="Delete" onClick={() => deletePlace(place.id)} />
-        </div>
+        {#if isAuthorizedToWrite}
+          <div class="actions">
+            <Button label="Edit" onClick={() => editPlace(place)} />
+            <Button label="Delete" onClick={() => deletePlace(place.id)} />
+          </div>
+        {/if}
       {/if}
     </Card>
   {/each}
